categorization tab as a separate module
